chore: remove simple- prefix from names
- remove unused files and tests
