+++ packages/ui/src/components/heading.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { dashboardCatalog } from "../lib/catalog.js";
-import type { CatalogueComponentProps } from "../lib/simple-component-catalog.js";
+import type { CatalogueComponentProps } from "../lib/component-catalog.js";
 
 export function Heading({
   element,
