+++ packages/ui/src/lib/renderer.test.tsx
@@ -5,14 +5,46 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import { createElement, type ReactNode } from "react";
 import { renderToStaticMarkup } from "react-dom/server";
 import { render, screen, waitFor, act } from "@testing-library/react";
-import {
-  Renderer,
-  type ComponentRenderProps,
-  type ComponentRegistry,
-} from "./renderer.js";
+import { Renderer, type RendererComponentProps } from "./renderer.js";
+import { createRegistry } from "./create-registry.js";
 import { KopaiSDKProvider } from "./kopai-provider.js";
-import type { UITree } from "./dynamic-component-catalog.js";
+import { createCatalog } from "./component-catalog.js";
+import z from "zod";
 import type { KopaiClient } from "@kopai/sdk";
+
+// Create a simple catalog and derive UITree type
+const _testCatalog = createCatalog({
+  name: "test",
+  components: {
+    Box: {
+      hasChildren: true,
+      description: "A box",
+      props: z.object({}),
+    },
+    Text: {
+      hasChildren: false,
+      description: "Text",
+      props: z.object({ content: z.string() }),
+    },
+    Capture: {
+      hasChildren: false,
+      description: "Captures props",
+      props: z.object({ content: z.string() }),
+    },
+    DataComponent: {
+      hasChildren: false,
+      description: "Data test component",
+      props: z.object({}),
+    },
+    RefetchComponent: {
+      hasChildren: false,
+      description: "Refetch test component",
+      props: z.object({}),
+    },
+  },
+});
+
+type UITree = z.infer<typeof _testCatalog.uiTreeSchema>;
 
 type MockClient = {
   searchTracesPage: ReturnType<typeof vi.fn>;
@@ -35,7 +67,10 @@ function createWrapper(client: MockClient) {
 }
 
 // Simple test components
-function Box({ element, children }: ComponentRenderProps) {
+function Box({
+  element,
+  children,
+}: RendererComponentProps<typeof _testCatalog.components.Box>) {
   return createElement(
     "div",
     { "data-type": element.type, "data-key": element.key },
@@ -43,12 +78,51 @@ function Box({ element, children }: ComponentRenderProps) {
   );
 }
 
-function Text({ element }: ComponentRenderProps) {
-  const { content } = element.props as { content: string };
+function Text({
+  element,
+}: RendererComponentProps<typeof _testCatalog.components.Text>) {
+  const { content } = element.props;
   return createElement("span", null, content);
 }
 
-const registry: ComponentRegistry = { Box, Text };
+function Capture(
+  _props: RendererComponentProps<typeof _testCatalog.components.Capture>
+) {
+  return createElement("div", null, "captured");
+}
+
+function DataComponent(
+  props: RendererComponentProps<typeof _testCatalog.components.DataComponent>
+) {
+  if (!props.hasData) {
+    return createElement("div", { "data-testid": "no-data" }, "No data source");
+  }
+  const { data, loading, error } = props;
+  if (loading)
+    return createElement("div", { "data-testid": "loading" }, "Loading...");
+  if (error)
+    return createElement("div", { "data-testid": "error" }, error.message);
+  return createElement("div", { "data-testid": "data" }, JSON.stringify(data));
+}
+
+function RefetchComponent(
+  props: RendererComponentProps<typeof _testCatalog.components.RefetchComponent>
+) {
+  if (!props.hasData) return null;
+  return createElement(
+    "div",
+    { "data-testid": "data" },
+    JSON.stringify(props.data)
+  );
+}
+
+const registry = createRegistry(_testCatalog, {
+  Box,
+  Text,
+  Capture,
+  DataComponent,
+  RefetchComponent,
+});
 
 describe("Renderer", () => {
   it("renders null for null tree", () => {
@@ -59,7 +133,7 @@ describe("Renderer", () => {
   });
 
   it("renders null for tree without root", () => {
-    const tree = { root: "", elements: {} } as UITree;
+    const tree = { root: "", elements: {} } as unknown as UITree;
     const result = renderToStaticMarkup(
       createElement(Renderer, { tree, registry })
     );
@@ -67,12 +141,18 @@ describe("Renderer", () => {
   });
 
   it("renders single element", () => {
-    const tree: UITree = {
+    const tree = {
       root: "text-1",
       elements: {
-        "text-1": { key: "text-1", type: "Text", props: { content: "Hello" } },
+        "text-1": {
+          key: "text-1",
+          type: "Text",
+          children: [],
+          parentKey: "",
+          props: { content: "Hello" },
+        },
       },
-    };
+    } satisfies UITree; // should be like this, not casts
     const result = renderToStaticMarkup(
       createElement(Renderer, { tree, registry })
     );
@@ -80,7 +160,7 @@ describe("Renderer", () => {
   });
 
   it("renders nested elements", () => {
-    const tree: UITree = {
+    const tree = {
       root: "box-1",
       elements: {
         "box-1": {
@@ -88,15 +168,17 @@ describe("Renderer", () => {
           type: "Box",
           props: {},
           children: ["text-1"],
+          parentKey: "",
         },
         "text-1": {
           key: "text-1",
           type: "Text",
           props: { content: "Nested" },
+          children: [],
           parentKey: "box-1",
         },
       },
-    };
+    } satisfies UITree;
     const result = renderToStaticMarkup(
       createElement(Renderer, { tree, registry })
     );
@@ -106,10 +188,16 @@ describe("Renderer", () => {
   });
 
   it("renders deeply nested tree", () => {
-    const tree: UITree = {
+    const tree = {
       root: "box-1",
       elements: {
-        "box-1": { key: "box-1", type: "Box", props: {}, children: ["box-2"] },
+        "box-1": {
+          key: "box-1",
+          type: "Box",
+          props: {},
+          children: ["box-2"],
+          parentKey: "",
+        },
         "box-2": {
           key: "box-2",
           type: "Box",
@@ -121,10 +209,11 @@ describe("Renderer", () => {
           key: "text-1",
           type: "Text",
           props: { content: "Deep" },
+          children: [],
           parentKey: "box-2",
         },
       },
-    };
+    } satisfies UITree;
     const result = renderToStaticMarkup(
       createElement(Renderer, { tree, registry })
     );
@@ -132,42 +221,8 @@ describe("Renderer", () => {
     expect(result).toContain('data-key="box-2"');
   });
 
-  it("warns for unknown component type", () => {
-    const warn = vi.spyOn(console, "warn").mockImplementation(() => {});
-    const tree: UITree = {
-      root: "unknown-1",
-      elements: {
-        "unknown-1": { key: "unknown-1", type: "Unknown", props: {} },
-      },
-    };
-    const result = renderToStaticMarkup(
-      createElement(Renderer, { tree, registry })
-    );
-    expect(result).toBe("");
-    expect(warn).toHaveBeenCalledWith(
-      "No renderer for component type: Unknown"
-    );
-    warn.mockRestore();
-  });
-
-  it("uses fallback for unknown component type", () => {
-    function Fallback({ element }: ComponentRenderProps) {
-      return createElement("div", { "data-fallback": element.type });
-    }
-    const tree: UITree = {
-      root: "unknown-1",
-      elements: {
-        "unknown-1": { key: "unknown-1", type: "Missing", props: {} },
-      },
-    };
-    const result = renderToStaticMarkup(
-      createElement(Renderer, { tree, registry, fallback: Fallback })
-    );
-    expect(result).toBe('<div data-fallback="Missing"></div>');
-  });
-
   it("skips children with missing elements", () => {
-    const tree: UITree = {
+    const tree = {
       root: "box-1",
       elements: {
         "box-1": {
@@ -175,14 +230,17 @@ describe("Renderer", () => {
           type: "Box",
           props: {},
           children: ["missing-1", "text-1"],
+          parentKey: "",
         },
         "text-1": {
           key: "text-1",
           type: "Text",
           props: { content: "Present" },
+          children: [],
+          parentKey: "box-1",
         },
       },
-    };
+    } satisfies UITree;
     const result = renderToStaticMarkup(
       createElement(Renderer, { tree, registry })
     );
@@ -191,27 +249,40 @@ describe("Renderer", () => {
   });
 
   it("passes hasData=false for elements without dataSource", () => {
-    let receivedProps: ComponentRenderProps | null = null;
-    function Capture(props: ComponentRenderProps) {
+    let receivedProps: RendererComponentProps<
+      typeof _testCatalog.components.Capture
+    > | null = null;
+    function CaptureLocal(
+      props: RendererComponentProps<typeof _testCatalog.components.Capture>
+    ) {
       receivedProps = props;
       return createElement("div", null, "captured");
     }
-    const tree: UITree = {
+    const localRegistry = createRegistry(_testCatalog, {
+      Box,
+      Text,
+      Capture: CaptureLocal,
+      DataComponent,
+      RefetchComponent,
+    });
+    const tree = {
       root: "capture-1",
       elements: {
         "capture-1": {
           key: "capture-1",
           type: "Capture",
-          props: { foo: "bar" },
+          props: { content: "hello" },
+          children: [],
+          parentKey: "",
         },
       },
-    };
+    } satisfies UITree;
     renderToStaticMarkup(
-      createElement(Renderer, { tree, registry: { Capture } })
+      createElement(Renderer, { tree, registry: localRegistry })
     );
     expect(receivedProps).not.toBeNull();
     expect(receivedProps!.hasData).toBe(false);
-    expect(receivedProps!.element.props).toEqual({ foo: "bar" });
+    expect(receivedProps!.element.props).toEqual({ content: "hello" });
   });
 });
 
@@ -234,47 +305,27 @@ describe("Renderer with dataSource", () => {
     vi.clearAllMocks();
   });
 
-  function DataComponent(props: ComponentRenderProps) {
-    if (!props.hasData) {
-      return createElement(
-        "div",
-        { "data-testid": "no-data" },
-        "No data source"
-      );
-    }
-    const { data, loading, error } = props;
-    if (loading)
-      return createElement("div", { "data-testid": "loading" }, "Loading...");
-    if (error)
-      return createElement("div", { "data-testid": "error" }, error.message);
-    return createElement(
-      "div",
-      { "data-testid": "data" },
-      JSON.stringify(data)
-    );
-  }
-
-  const dataRegistry: ComponentRegistry = { DataComponent };
-
   it("passes data props to component with dataSource", async () => {
     mockClient.searchTracesPage.mockResolvedValueOnce({
       data: [{ traceId: "abc" }],
     });
 
-    const tree: UITree = {
+    const tree = {
       root: "data-1",
       elements: {
         "data-1": {
           key: "data-1",
           type: "DataComponent",
           props: {},
+          children: [],
+          parentKey: "",
           dataSource: { method: "searchTracesPage", params: { limit: 10 } },
         },
       },
-    };
+    } satisfies UITree;
 
     const Wrapper = createWrapper(mockClient);
-    render(createElement(Renderer, { tree, registry: dataRegistry }), {
+    render(createElement(Renderer, { tree, registry }), {
       wrapper: Wrapper,
     });
 
@@ -297,20 +348,22 @@ describe("Renderer with dataSource", () => {
     });
     mockClient.searchTracesPage.mockReturnValueOnce(promise);
 
-    const tree: UITree = {
+    const tree = {
       root: "data-1",
       elements: {
         "data-1": {
           key: "data-1",
           type: "DataComponent",
           props: {},
+          children: [],
+          parentKey: "",
           dataSource: { method: "searchTracesPage", params: {} },
         },
       },
-    };
+    } satisfies UITree;
 
     const Wrapper = createWrapper(mockClient);
-    render(createElement(Renderer, { tree, registry: dataRegistry }), {
+    render(createElement(Renderer, { tree, registry }), {
       wrapper: Wrapper,
     });
 
@@ -327,20 +380,22 @@ describe("Renderer with dataSource", () => {
       new Error("Network error")
     );
 
-    const tree: UITree = {
+    const tree = {
       root: "data-1",
       elements: {
         "data-1": {
           key: "data-1",
           type: "DataComponent",
           props: {},
+          children: [],
+          parentKey: "",
           dataSource: { method: "searchTracesPage", params: {} },
         },
       },
-    };
+    } satisfies UITree;
 
     const Wrapper = createWrapper(mockClient);
-    render(createElement(Renderer, { tree, registry: dataRegistry }), {
+    render(createElement(Renderer, { tree, registry }), {
       wrapper: Wrapper,
     });
 
@@ -357,7 +412,11 @@ describe("Renderer with dataSource", () => {
 
     let capturedRefetch: ((params?: Record<string, unknown>) => void) | null =
       null;
-    function RefetchComponent(props: ComponentRenderProps) {
+    function RefetchComponentLocal(
+      props: RendererComponentProps<
+        typeof _testCatalog.components.RefetchComponent
+      >
+    ) {
       if (!props.hasData) return null;
       capturedRefetch = props.refetch;
       return createElement(
@@ -367,20 +426,30 @@ describe("Renderer with dataSource", () => {
       );
     }
 
-    const tree: UITree = {
+    const localRegistry = createRegistry(_testCatalog, {
+      Box,
+      Text,
+      Capture,
+      DataComponent,
+      RefetchComponent: RefetchComponentLocal,
+    });
+
+    const tree = {
       root: "data-1",
       elements: {
         "data-1": {
           key: "data-1",
           type: "RefetchComponent",
           props: {},
+          children: [],
+          parentKey: "",
           dataSource: { method: "searchTracesPage", params: {} },
         },
       },
-    };
+    } satisfies UITree;
 
     const Wrapper = createWrapper(mockClient);
-    render(createElement(Renderer, { tree, registry: { RefetchComponent } }), {
+    render(createElement(Renderer, { tree, registry: localRegistry }), {
       wrapper: Wrapper,
     });
 
@@ -398,24 +467,124 @@ describe("Renderer with dataSource", () => {
   });
 
   it("renders element without dataSource normally", () => {
-    const tree: UITree = {
+    const tree = {
       root: "data-1",
       elements: {
         "data-1": {
           key: "data-1",
           type: "DataComponent",
           props: {},
+          children: [],
+          parentKey: "",
           // No dataSource
         },
       },
-    };
+    } satisfies UITree;
 
     const Wrapper = createWrapper(mockClient);
-    render(createElement(Renderer, { tree, registry: dataRegistry }), {
+    render(createElement(Renderer, { tree, registry }), {
       wrapper: Wrapper,
     });
 
     expect(screen.getByTestId("no-data")).toBeDefined();
     expect(screen.getByTestId("no-data").textContent).toBe("No data source");
+  });
+});
+
+describe("createRegistry with Renderer integration", () => {
+  const integrationCatalog = createCatalog({
+    name: "integration-test",
+    components: {
+      Wrapper: {
+        hasChildren: true,
+        description: "A wrapper",
+        props: z.object({}),
+      },
+      Label: {
+        hasChildren: false,
+        description: "A label",
+        props: z.object({ text: z.string() }),
+      },
+    },
+  });
+
+  type IntegrationUITree = z.infer<typeof integrationCatalog.uiTreeSchema>;
+
+  function Wrapper({
+    children,
+  }: RendererComponentProps<typeof integrationCatalog.components.Wrapper>) {
+    return createElement("div", { "data-testid": "wrapper" }, children);
+  }
+
+  function Label({
+    element,
+  }: RendererComponentProps<typeof integrationCatalog.components.Label>) {
+    return createElement(
+      "span",
+      { "data-testid": "label" },
+      element.props.text
+    );
+  }
+
+  it("renders tree using registry created with createRegistry", () => {
+    const registry = createRegistry(integrationCatalog, {
+      Wrapper,
+      Label,
+    });
+
+    const tree: IntegrationUITree = {
+      root: "wrapper-1",
+      elements: {
+        "wrapper-1": {
+          key: "wrapper-1",
+          type: "Wrapper",
+          props: {},
+          children: ["label-1"],
+          parentKey: "",
+        },
+        "label-1": {
+          key: "label-1",
+          type: "Label",
+          props: { text: "Hello World" },
+          children: [],
+          parentKey: "wrapper-1",
+        },
+      },
+    };
+
+    const result = renderToStaticMarkup(
+      createElement(Renderer, { tree, registry })
+    );
+
+    expect(result).toContain("Hello World");
+    expect(result).toContain('data-testid="wrapper"');
+    expect(result).toContain('data-testid="label"');
+  });
+
+  it("accepts registry from createRegistry without cast", () => {
+    const registry = createRegistry(integrationCatalog, {
+      Wrapper,
+      Label,
+    });
+
+    const tree: IntegrationUITree = {
+      root: "label-1",
+      elements: {
+        "label-1": {
+          key: "label-1",
+          type: "Label",
+          props: { text: "Test" },
+          children: [],
+          parentKey: "",
+        },
+      },
+    };
+
+    // This should compile without any cast - Renderer is now generic
+    const result = renderToStaticMarkup(
+      createElement(Renderer, { tree, registry })
+    );
+
+    expect(result).toContain("Test");
   });
 });
